+++ src/main/java/pl/smsapi/api/ContactsFactory.java
@@ -2,13 +2,18 @@ package pl.smsapi.api;
 
 import pl.smsapi.Client;
 import pl.smsapi.api.action.contacts.*;
+import pl.smsapi.api.action.contacts.groups.GroupsFactory;
 import pl.smsapi.proxy.Proxy;
 import pl.smsapi.proxy.ProxyNative;
 
+/**
+ * @deprecated use {@link pl.smsapi.api.action.contacts.ContactsFactory()} or {@link pl.smsapi.api.action.contacts.groups.GroupsFactory()} instead
+ */
+@Deprecated
 public class ContactsFactory extends ActionFactory {
 
     /**
-     * @deprecated use {@link ContactsFactory(Client, Proxy)} instead
+     * @deprecated use {@link ContactsFactory(Client, Proxy)} or {@link pl.smsapi.api.action.contacts.ContactsFactory(Client, Proxy)} instead
      */
     @Deprecated
     public ContactsFactory(Client client) {
@@ -20,6 +25,10 @@ public class ContactsFactory extends ActionFactory {
         super(client, proxy);
     }
 
+    /**
+     * @deprecated use {@link pl.smsapi.api.action.contacts.ContactsFactory#actionList()} instead
+     */
+    @Deprecated
     public ContactsContactList actionContactList() {
         ContactsContactList action = new ContactsContactList();
         action.client(client);
@@ -27,6 +36,10 @@ public class ContactsFactory extends ActionFactory {
         return action;
     }
 
+    /**
+     * @deprecated use {@link pl.smsapi.api.action.contacts.ContactsFactory#actionGet(String)} instead
+     */
+    @Deprecated
     public ContactsContactGet actionContactGet(String contactId) {
         ContactsContactGet action = new ContactsContactGet();
         action.client(client);
@@ -35,6 +48,10 @@ public class ContactsFactory extends ActionFactory {
         return action;
     }
 
+    /**
+     * @deprecated use {@link pl.smsapi.api.action.contacts.ContactsFactory#actionAdd()} instead
+     */
+    @Deprecated
     public ContactsContactAdd actionContactAdd() {
         ContactsContactAdd action = new ContactsContactAdd();
         action.client(client);
@@ -42,6 +59,10 @@ public class ContactsFactory extends ActionFactory {
         return action;
     }
 
+    /**
+     * @deprecated use {@link pl.smsapi.api.action.contacts.ContactsFactory#actionDelete(String)} instead
+     */
+    @Deprecated
     public ContactsContactDelete actionContactDelete(String contactId) {
         ContactsContactDelete action = new ContactsContactDelete();
         action.client(client);
@@ -50,6 +71,10 @@ public class ContactsFactory extends ActionFactory {
         return action;
     }
 
+    /**
+     * @deprecated use {@link pl.smsapi.api.action.contacts.ContactsFactory#actionEdit(String)} instead
+     */
+    @Deprecated
     public ContactsContactEdit actionContactEdit(String contactId) {
         ContactsContactEdit action = new ContactsContactEdit();
         action.client(client);
@@ -58,6 +83,10 @@ public class ContactsFactory extends ActionFactory {
         return action;
     }
 
+    /**
+     * @deprecated use {@link GroupsFactory#actionList()} instead
+     */
+    @Deprecated
     public ContactsGroupList actionGroupList() {
         ContactsGroupList action = new ContactsGroupList();
         action.client(client);
@@ -65,6 +94,10 @@ public class ContactsFactory extends ActionFactory {
         return action;
     }
 
+    /**
+     * @deprecated use {@link GroupsFactory#actionAdd(String)} ()} instead
+     */
+    @Deprecated
     public ContactsGroupAdd actionGroupAdd() {
         ContactsGroupAdd action = new ContactsGroupAdd();
         action.client(client);
@@ -72,6 +105,10 @@ public class ContactsFactory extends ActionFactory {
         return action;
     }
 
+    /**
+     * @deprecated use {@link GroupsFactory#actionDelete(String)} ()} instead
+     */
+    @Deprecated
     public ContactsGroupDelete actionGroupDelete(String groupId) {
         ContactsGroupDelete action = new ContactsGroupDelete();
         action.client(client);
@@ -80,6 +117,10 @@ public class ContactsFactory extends ActionFactory {
         return action;
     }
 
+    /**
+     * @deprecated use {@link GroupsFactory#actionGet(String)} ()} instead
+     */
+    @Deprecated
     public ContactsGroupGet actionGroupGet(String groupId) {
         ContactsGroupGet action = new ContactsGroupGet();
         action.client(client);
@@ -88,6 +129,10 @@ public class ContactsFactory extends ActionFactory {
         return action;
     }
 
+    /**
+     * @deprecated use {@link GroupsFactory#actionEdit(String)} ()} instead
+     */
+    @Deprecated
     public ContactsGroupEdit actionGroupEdit(String groupId) {
         ContactsGroupEdit action = new ContactsGroupEdit();
         action.client(client);