+++ src/main/java/pl/smsapi/api/action/contacts/ContactsContactAdd.java
@@ -1,5 +1,9 @@
 package pl.smsapi.api.action.contacts;
 
+/**
+ * @deprecated use {@link ContactAdd} instead
+ */
+@Deprecated
 public class ContactsContactAdd extends AbstractContactsContactControl<ContactsContactAdd> {
 
     @Override