+++ src/main/java/pl/smsapi/api/action/contacts/ContactsGroupAdd.java
@@ -1,5 +1,11 @@
 package pl.smsapi.api.action.contacts;
 
+import pl.smsapi.api.action.contacts.groups.GroupAdd;
+
+/**
+ * @deprecated use {@link GroupAdd} instead
+ */
+@Deprecated
 public class ContactsGroupAdd extends AbstractContactsGroupControl<ContactsGroupAdd> {
 
     @Override