+++ src/main/java/pl/smsapi/api/response/contacts/ContactsContactListResponse.java
@@ -4,6 +4,10 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 import pl.smsapi.api.response.ListResponse;
 
+/**
+ * @deprecated use {@link pl.smsapi.api.action.contacts.Contacts()} instead
+ */
+@Deprecated
 public class ContactsContactListResponse extends ListResponse<ContactsContactResponse> {
 
     public ContactsContactListResponse(int count, JSONArray jsonArray) {